Procfile

+++ app/Http/Controllers/LocationsController.php
@@ -7,18 +7,21 @@ use App\Location;
 
 class LocationsController extends Controller
 {
-    // getでlocations/にアクセスされた場合の「一覧表示処理」
-    public function index()
+     public function index()
     {
-        $locations = Location::all();
-        //dd($locations); //追加
-        return view('locations.index', [
-            'locations' => $locations,
-        ]);
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $locations = Location::orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'locations' => $locations,
+            ];
+        }
+        return view('locations.index', $data);
+      
     }
-
-   
-   // getでlocations/createにアクセスされた場合の「新規登録画面表示処理」
+    
     public function create()
     {
         $location = new Location;
@@ -27,48 +30,34 @@ class LocationsController extends Controller
             'location' => $location,
         ]);
     }
-     
-    // postでlocations/にアクセスされた場合の「新規登録処理」
-    public function store(Request $request)
+    
+     public function store(Request $request)
     {
         $this->validate($request, [
-            'content' => 'required|max:140',
             'place' => 'required|max:50',
-            'area' => 'required',
-            'thefile' => [
-               'required',
-               'file',
-               'image',
-               'mimes:jpeg,png',
-            ]
+            'facility' => 'required',
         ]);
-    
-        //画像ファイル受け取り処理
-        $filename='';
-        if ($request->file('thefile')->isValid([])) {
-            $filename = $request->file('thefile')->store('img');
+        
+        $location = $this->getLocation($request->place);
+        
+        if($location == null){
+            return back()->with('error', '指定された場所が存在しません。');
         }
-        
-        //時間のセット
-        date_default_timezone_set('Asia/Tokyo');
-        $now = date("Y/m/d H:i");
-        
-        $location = new Location;
-        $location->content = $request->content;
-        $location->area = $request->area;
-        $location->place = $request->place;
-        $location->time = $now;
-        $location->image = $filename;
-        $location->save();
-    
-        
-        return redirect('/');
+        else{
+            $lat = $location['lat']; // latを取得
+            $lng = $location['lng']; // lngを取得
+            $request->user()->locations()->create([
+                'facility'=> $request->facility,
+                'place' => $request->place,
+                'lat' => $lat,
+                'lng' => $lng,
+            ]);
+            
+            return redirect('/locations');
+        }
     }
     
-   
-     
-   // getでlocations/idにアクセスされた場合の「取得表示処理」
-    public function show($id)
+     public function show($id)
     {
         $location = Location::find($id);
 
@@ -76,64 +65,18 @@ class LocationsController extends Controller
             'location' => $location,
         ]);
     }
-
     
-     // getでlocations/id/editにアクセスされた場合の「更新画面表示処理」
-    public function edit($id)
-    {
-        $location = Location::find($id);
-
-        return view('locations.edit', [
-            'location' => $location,
-        ]);
-    }
-
-  
-     
-   // putまたはpatchs/idにアクセスされた場合の「更新処理」
-    public function update(Request $request, $id)
-    {
-        $this->validate($request, [
-            'content' => 'required|max:140',
-            'place' => 'required|max:50',
-            'area' => 'required',
-            'thefile' => [
-                'required',
-                'file',
-                'image',
-                'mimes:jpeg,png',
-            ]
-        ]);
-    
-        //画像ファイル受け取り処理
-        $filename='';
-        if ($request->file('thefile')->isValid([])) {
-            $filename = $request->file('thefile')->store('img');
-        }
-        
-        //時間のセット
-        date_default_timezone_set('Asia/Tokyo');
-        $now = date("Y/m/d H:i");
-        
-        //$location = new Location::find($id);
-        $location = Location::find($id);
-        $location->content = $request->content;
-        $location->area = $request->area;
-        $location->place = $request->place;
-        $location->time = $now;
-        $location->image = $filename;
-        $location->save();
-        
-        return redirect('/');
-    }
-    
-    
-    // deleteでlocations/idにアクセスされた場合の「削除処理」
     public function destroy($id)
     {
-        $location = Location::find($id);
-        $location->delete();
-
-        return redirect('/');
+        $location = \App\Location::find($id);
+        if (\Auth::id() === $location->user_id) {
+            $location->delete();
+        }
+        if(url()->previous() == "/users/{user}"){
+            return redirect('/locations');
+        }
+        else{
+            return back();
+        }
     }
 }
